Refactoring [skip ci]
